backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
sqlmodel==0.0.22
sqlalchemy==2.0.35
alembic==1.13.3
httpx>=0.28.1,<1.0.0
python-dotenv==1.0.1
PyMuPDF==1.24.9
google-genai==1.3.0
feedparser==6.0.11
faiss-cpu==1.8.0.post1
